feat: :art: Seperated Map concerns from Main
Took out all the Map related functions, as well as the toggle funcction, and added to the Map.ts and DOMManipulation.ts files.
